Add crosstable function
This function is a generic cross-tabulation function like 
`table`, but it contains options to weight the data and 
to place marginals on the sides. It also labels the 
columns descriptively and gives an option to present
the table as percentages. And because the output is a data_frame, it can be wrapped with kable() for print-quality output.

This was migrated from the (somewhat stale) nhtsHelper repository.