Only convert to factor if not already
If the user gives a numeric or integer value as `color_field`, ggplot2 will only work properly if we convert that field to a color. #22 pointed out that if a variable is already a factor, it might be best to leave it alone.

This introduces a pass such that variables are only factored if they are not already factors.